AssetCheckSummaryRecord (#21940)
Adds an "AssetCheckState" API; which functions similarly to an
AssetRecord. The core conceit is to be able to query just based on asset
check key a set of information about the "state" of the asset check
(most recent run, for now), without having any references to the event
log.

For now, it's actually powered by the event log, however.

How I tested this
- Added a new semi-end-to-end test to our event log storage tests. This
required me setting up an instance hooked up to an event log storage;
which I think is nice to have anyway.
- There's one failing test here that's confusing me;
test_get_event_records_sqlite. I changed it to use the instance passed
in by fixture, which should in theory not change the results? But it
seems like it has. Probably missing something here.

There will be a follow up cloud pr.