[4/5] Run Timeline IndexedDB Caching (#22009)
## Summary & Motivation

Store the Run Timeline data in indexedb so that new page loads can reuse
data from previous page loads.

Q: How is data cleared?
A: Whenever we load the page and create the cache instance we check if
the cache has entries over a week old and delete them.

Q: How is data keyed in OSS?
A: We use the run_storage_id which is unique to the instance. That way
if a user blows away their instance and start fresh they will get a
completely separate cache.

Q: How is data keyed in Cloud?
A:  https://github.com/dagster-io/internal/pull/9828

## How I Tested These Changes

Extensive use of the RunTimeline in cache + jest.