Fix sensor/schedule state in the left nav being outdated (#22032)
## Summary & Motivation

In a previous PR, we stopped storing the RootWorkspaceQuery in the
Apollo cache. However, this change affected the left navigation, which
relied on data from the workspace context. As a result, the left
navigation stopped receiving updates to the schedule/sensor state after
mutations, since its data source was no longer tied to the Apollo cache.

To resolve this issue, I updated the left navigation to query for the
latest status separately. This new query is stored in the Apollo cache,
ensuring that any changes to the state on the sensor page or the left
navigation are reflected in both places.

## How I Tested These Changes



https://github.com/dagster-io/dagster/assets/2286579/39200fb1-ed51-4620-b725-66b3893d1c02