[dagster-dbt] force duckdb adapter to use abs path, fixing flaky test (#22042)
## Summary

In order to use the DuckDB dbt adapter for postprocessing tasks like
querying row count, we needed to ensure the db filepath resolved
properly. Often, this filepath was relative to the dbt project
directory, requiring us to use `os.pushd` so that accesses would happen
with the dbt project dir as working directory.

Unfortunately, this working dir was occasionally getting overridden in
between opening the `pushd` context and actually executing queries. I'm
not sure what exactly was responsible for adjusting the working
directory, but the status quo is pretty fragile.

Under the hood, this PR resorts to updating duckdb db paths in the dbt
adapter to be absolute paths. This makes the test run much more reliably
locally, and means we no longer need to set the working directory.

Also cleans up some of the duckdb-specific logic related to iterators.

## Test Plan

Existing tests, ran locally many times to ensure not flaky