Fix fast partition count calculations for daily partitions definition
Summary:
owen tried to warn me about this on https://github.com/dagster-io/dagster/pull/21791/files, and in my hubris I did not listen - the test coverage here was not as thorough as i thought it was. Fix the fast calculations when you have a simple daily partitions definition (the existing logic gets messed up by DST boundaries when they are on opposite sides of the boundary).

Test Plan: BK, shadow graphql of an incorrect partitions count