Introduce `StorageKindTagSet` (#22029)
## Summary

Introduces `StorageKindTagSet` tag collection which includes the `dagster/storage_type` asset tag. This tag can be used to describe where an asset is actually stored. We can use this in stacked PRs to specially display this storage information in the UI specially as well as automatically populating the tag in our various integrations.

## Test Plan

Unit tests.