[TWP][2/n] Remove PartitionKeysTimeWindowPartitionSubset (#22119)
## Summary & Motivation

This class was created to make up for the fact that we had no efficient way of doing basic set operations efficiently on time-window-based subsets. We've since fixed these issues, and now the behavior of by, default, calling `my_time_partitions_def.subset_with_partition_keys(...)` returning a PartitionKeysTimeWindowPartitionsSubset is a performance liability, as it means that it will result in slow operations when used against TimeWindowPartitionsSubsets. Removing this class entirely greatly simplifies the mental model here.

Note that this means we can also remove the BaseTimeWindowPartitionsSubset class, I just wanted to keep this PR a bit smaller, so I'll do that in a follow up.


## How I Tested These Changes