[ui] Show materialization status for External Assets (#22124)
## Summary & Motivation

This PR is a fix for FE-365
(https://linear.app/dagster-labs/issue/FE-365/report-event-is-creating-materialization-events-on-external-assets)
and reflects this goal:

Assets with isObservable=true should show "Last Observation", and allow
you to click "Observe". All other assets should show materialization
state.

To date, we've used `isSource` in a lot of these places when we meant
`isObservable`.

Note: Work here is likely not done. We still use `isSource` in places
where we mean `isExecutable`, and there are places in the UI (screenshot
below) we refer to "Source asset". We may need to take a tour through
the remaining callsites on a quiet(er) day.

## How I Tested These Changes

I created a source asset (below) with a materialization, and verified
that I now see the materialization oriented stats. I also verified that
for `observable_source_assets`, I still see "Last Observation", etc.


![image](https://github.com/dagster-io/dagster/assets/1037212/95b9d0b9-e806-4bc7-835a-de32dd1a26a8)

![image](https://github.com/dagster-io/dagster/assets/1037212/37e22506-7807-430e-99b6-ab1b7bb18535)

---------

Co-authored-by: bengotow <bgotow@elementl.com>