Fix RunTimeline bucket pagination (#22131)
## Summary & Motivation

Currently when we're paginating we call `getMissingIntervals` within the
paginated fetch loop. This is a bug because after the first fetch we add
data for the time range and then `getMissingIntervals` ends up returning
a 1 millisecond time range for the subsequent paginated fetches of that
bucket. Instead we should be calling getMissingIntervals when
determining the bucket to fetch and then when paginating we keep using
the same bucket.

## How I Tested These Changes

Loaded a timeline page, set the BATCH_LIMIT to 1 and made sure we
paginated and fetched the rest of the runs for the bucket (current bug
would result in just 1 run per bucket).

Wrote a jest test