RFC: What if we didn't say op (#22190)
## Summary & Motivation

Someone without intimate knowledge of Dagster internals may not always
understand what an `op` is and, I would argue, does not always need to
understand it.

In this example, I replace `op` with `operation` but could also use
words like `computation`. The idea is to communicate to the user what is
happening without introducing a concept they may not necessarily need to
know about.