`ShellCommandBlueprint` and `BlueprintAssetsDefinition` (#22104)
## Summary & Motivation

This PR explores a few related ideas:
- Out-of-the-box blueprints corresponding to standard pipes clients
(databricks & subprocess).
- Single-assets and multi-asset blueprints as separate blueprints.
- Base blueprint types for helping define out-of-the-box blueprints

One question that came up when implementing multi-asset blueprints was
what to name the equivalent of `AssetSpec` in blueprint-land. Which
points to a general question of what to name classes that are part of
defining blueprints but don't correspond to code location-level
definitions of their own.

This PR goes with `AssetSpecModel`. An alternative would have been
`AssetSpecBlueprint`. However, the `Blueprint` base class currently has
a `build_defs` method, which wouldn't make sense for
`AssetSpecBlueprint`. We could potentially make this work by renaming
the `Blueprint` base class to `DefinitionsBlueprint`.

## How I Tested These Changes