Stop hiding server errors (#22114)
This will allow the user to get the relevant information on the real
error that caused the exception (such as 403)

## Summary & Motivation
We encountered a GraphQL error and spent a few hours debugging it only
to find out it was an 403 error. Sadly, this message was hidden by the
general except clause. We believe this change may be helpful for other
users.

## How I Tested These Changes
Queried to get an 403 error on our servers and see the updated error
message.