[DS][46/n] Create NewlyUpdatedCondition (#21973)
## Summary & Motivation

As title. When testing this, I realized that two latent bugs were impacting the correctness here.

Essentially, when computing the set of things changed since a given cursor inside the CachingInstanceQueryer, we return a new cursor representing the latest storage id across all records that were used to make the decision (i.e. the target asset's max storage id and all of its parents max storage ids).

However, we were shortcutting this logic in the case that a child had no parent assets, meaning an asset with no parents would never have its cursor advance!

After fixing that, I noticed that a separate test started failing, dealing with how we calculate which asset partitions are missing when you update the partitions definition.

The secondary bug (only present in the legacy missing rule, not the new version), happened because the legacy rule stores its knowledge of what's NOT missing inside the cursor, rather than just using the asset status cache (this is because it also wants to keep track of things that were previous requested as "not missing", something the new system handles in a more systematic way in this very stack).

When you update the partitions definition, this makes the AssetSubset get deserialized as an empty subset instead of None (again, this only happens in the legacy system), meaning the system now things there are zero partitions which are not missing!

However, the specific test happened to be managing to succeed despite this bug because the asset under test had no dependencies. So its cursor was always set to the beginning of time, and so it'd start off with this empty set of things that had been handled already, but then add in all of the partitions which had been materialized since the cursor (which was locked to "the beginning of time"), meaning it would actually accidentally arrive at the right answer, and my fix to make the cursor advance correctly broke the test.

So I fixed the secondary bug as well by just having special handling when a partitions definition changes (and to emphasize, this "treat an invalid partitions subset as an empty partitions subset" is terrible because of the possibility of these bugs and has been completely divested from in the AssetSlice world)

## How I Tested These Changes