[DS][52/n] Move prefetch step to Evaluator class (#22133)
## Summary & Motivation

This doesn't really change anything in the real world, but it ensures that our tests are on a level playing field with reality, as otherwise when you use the evaluator directly, you end up individually fetching each AssetRecord vs. fetching them in a large batch when using the AssetDaemon.

Alternative possibility would be to treat a big list of AssetRecords (and potentially AssetStatusCacheValues?) as inputs into the evaluator, although that is a larger change as we'd need to wire those into the CachingInstanceQueryer somehow. I think in reality we'd want to have that be a step on the path towards creating a serializable bucket of information that we use to construct an AssetGraphView with pre-loaded information across process boundaries.

## How I Tested These Changes