[DS][54/n] Make slice conversion more efficient (#22136)
## Summary & Motivation

It is not free to do isinstance checks, so I make two separate entrypoints for converting from asset subset to asset slice. For the common case, where we know that a subset is guaranteed to be valid, we skip the instance check entirely.

Also uses a cached_method for get_asset_slice, as the AssetSlice itself caches some properties/methods, and so returning the same object instance from get_asset_slice means different call sites get to share the same cache.

## How I Tested These Changes