[DS][56/n] Use slimmed cursor object for internal computation (#22157)
## Summary & Motivation

This is the first step in migrating us to using a slimmed-down cursor object within our internals. Currently, we store the entire AssetConditionEvaluationState object inside of our cursor. This object contains high-fidelity information about each node of computation in our graph. This _is_ necessary to store for use in the UI, as we want to record all relevant information about each node, but it is way more information than required for incremental calculations.

This PR splits off a simplified SchedulingConditionCursor class, which contains just a flattened mapping of condition node id to SchedulingConditionNodeCursor, which contains some relevant properties. We then refactor the internals to pull information off of this class. Because of this flattened structure, we have the freedom to store no information at all for nodes in the condition evaluation that do not require a cursor to function (which is the majority of them).

To keep things a bit simpler, this PR does not actually change any storage implementations (that'll happen in the next PR in the stack), and instead does a conversion from big complex object to simple cursor after deserializing the complex object from storage.

## How I Tested These Changes