[DS][57/n] Use slimmed cursor object in storage (#22158)
## Summary & Motivation

This is the continuation of the previous PR's work, where now we actually store the slimmed-down cursor in our serialized objects.

One decision I made here was to (at least temporarily) not delete any information in the old format from the cursor, which makes it a bit easier to roll back this change in the event of an emergency. We can always go back and wipe out the old value later.

On case in perf_test.py, this PR results in the cursor becoming almost 30x smaller.

## How I Tested These Changes