[DS][60/n] Remove parent_updated condition (#22189)
## Summary & Motivation

As title -- we want people to be using the "AnyDepsMatch(NewlyUpdated())" pattern instead of this. We may choose in the near to medium future to bring this back for the purposes of reconciliation-based scheduling (i.e. something that will always try to bring everything "up to date") but for now I think this confuses matters, and removing it ensures that when we bring it back we can determine exactly how we want to name it / model it etc.

## How I Tested These Changes