[DA] Fix error with freshness-based scheduling (#22286)
## Summary & Motivation

Long story short, this issue was lying in wait for a long time, but we never hit it because all of our BaseTimeWindowPartitionSubsets were in the "partition key" format rather than the "time window" format. If you check if <random non-time-window key, in the detected case, None> in <partition key set>, that's totally fine, but if you check if that same "invalid" key is in <time window set>, then you get an error.

This PR makes the time window set copy the same behavior of the non-time window set.

## How I Tested These Changes