Extract InOutMapper to begin refactoring AssetsDefinition construction process (#22222)
## Summary & Motivation

`multi_asset` is a beast, as well as the entire `AssetsDefinition` creation machinery.

This class introduces a class `InOutMapper` (renamed to `AssetsDefinitionFactory` upstack once its mandate increased) that begins to tease apart `multi_asset` which is nearly 300 LoC, has 37 local variables, and a huge dynamically scoped inner function.

To see where this is going I have created a digest [PR](https://github.com/dagster-io/dagster/pull/22238) that demonstrate the before after.

This also sets us up to consolidate the `AssetsDefinition` creation code paths, as it contains tons of duplicated code strewn about. Instead we will be able to invoke the decomposed code in the new factory functions.

This was motivated by the discussion in https://github.com/dagster-io/dagster/pull/22221 that suggested we move a propose class to be within the inheritance hierarchy of `AssetsDefinition`. The complexity of logic surrounding the construction of said object is completely out of control, and I found it effectively intractable to do an inheritance scheme cleanly.

## How I Tested These Changes

BK