optimize buildkite python packages change detection (#22141)
## Summary & Motivation

Determining buildkite pipeline steps locally is very slow (22s). I believe this is most likely due to __pycache__ directories or other similar files that are probably not an issue in CI. However, when iterating locally, it helps for this step to be as fast as possible, and these changes reduce local pipeline output from 22s to 2s. 

## How I Tested These Changes

- visually inspected that the same packages are found
- modified a python file in a package and ensured that it was still picked up as a changed package