Expunge freeze_pendulum_time / pendulum usage in tests, part 1
Summary:
First of a series in removal of pendulum from our tests - move freeze_pendulum_time over to freezegun instead (which coincidentally dosn't work well when you pass in a pendulum time to it, so that's a nice excuse to remove it from the tests that use it). The main beneficiary here is the scheduler tests.