Correctly determine the most recent eligible in progress run when computing in-progress status for an asset (#22306)
Summary:
The comment here indicates that it was intending to fetch the most
recent planned run, but the last_run_id column is updated on both
asset_materialization_planned and asset_materialization. So if e.g. a
backfill was started that materialized multiple runs at once, as soon as
a single one of those runs was updated, the in progress status would no
longer consider any of the other runs still in the queue planning to
materialize the asset.

Certain storages instead have the column that I believe we want here,
which is not updated when a materialization comes in, but instead only
when a materialization planned event comes in. Use that when possible,
otherwise fetched the planned event that we are looking for.

Test Plan:
Run a backfill targeting multiple partitions of the same asset at once -
see the run correctly marked as "in progress" until the last one is
materialized.

test change

more test change

## Summary & Motivation

## How I Tested These Changes