[3/n] Use workspace context for OverviewJobs/OverviewSchedules/OverviewSensors/PipelineJobs (#22390)
## Summary & Motivation

These pages are querying for data that is already in the
WorkspaceContext so we make them rely on the workspace context if its
available and their respective queries have not returned yet. The
workspace cache is currently including stateful data which allows it to
be used without changing types, but the stateful data might be out of
data so we only use it while their queries are loading. In a future PR I
will remove the stateful stuff from workspace context and remove the
definition stuff from the respective queries in these pages so that the
queries can avoid loading workspace snapshots but for now the backend
still fetches the workspace snapshots so deferring that to a future PR.


## How I Tested These Changes
Loaded each of these pages and confirmed they're using the
WorkspaceContext while their queries are loading