[ui] Modernize Description component, handle async rendering of Markdown (#22422)
Fixes
https://linear.app/dagster-labs/issue/FE-392/issue-with-cut-off-asset-descriptions-and-no-expandscroll-option

## Summary & Motivation

Was able to reliably reproduce this loading
http://localhost:3001/elementl/prod/assets/purina/cloud_reporting/reporting_deduped_internal_asset_materialization_metrics?view=overview&partition=2024-06-09-21%3A00.

It appears that this happens because our markdown support is
lazy-loaded, and if you navigate to this page and then hard-refresh, the
Description element renders empty on the first render, and then fills
with markdown content a few renders later. It's possible there's also
some React 18 deferred rendering stuff going on, but I'm less sure.

This component is old and less sophisticated than MiddleTruncate. I
updated it to use both a resize observer and a mutation observer (to
detect both the markdown render finishing and potentially the
description prop changing, which was also not supported before).

Co-authored-by: bengotow <bgotow@elementl.com>