Make AMP use most recent materialization OR observation for observable assets (#22350)
## Summary & Motivation

Fixes
https://linear.app/dagster-labs/issue/FOU-219/[schedulingcondition]-fix-handling-of-externalassets-with-the.

For observable external assets, AMP is unresponsive to reported asset
materializations because it only looks for the most recent asset
observation. This PR changes `CachingInstanceQueryer` to look for the
most recent materialization or observation for observables, which allows
AMP to respond to reported materializations.

## How I Tested These Changes

Modified the unit tests for the `NewlyUpdated` conditions to include an
observable asset.