Make compute_kind tag use prefix (#22359)
## Summary & Motivation

Currently the tag used to represent "compute kind" is unique among our
system tags in not being prefixed with our system tag prefix. This PR
fixes this situation by:

- Changing the tag used by all our internal APIs to
`<prefix>/compute_kind`
- Look for the legacy tag at OpDefinition construction time. If found,
print a warning and also assign the new tag. This leaves the definition
with both the legacy and new tags.
- Look for the legacy tag when deserializing `OpDefSnap`. If the new tag
is also present, do nothing (the snap was generated by a code server
recent enough to implement normalization at OpDefinition construction).
If the new tag is not also present, assign it on `OpDefSnap`. This
supports code servers too old to have the normalization.

I used 1.9.0 for the deprecation warning but maybe we should do 2.0.

## How I Tested These Changes

New unit tests.