[Perf monitoring] Don't update path using nesting route (#22434)
## Summary & Motivation

Some of our routes nest other routes. Eg: `/overview/activity` nests 3
other routes. Currently if a component rendering a nested route
re-renders it ends up changing the path back to its nested route path
(eg. `/overview/activity`) and if display-done happens at that point
then we end up using that as the view name. To avoid that just never use
the path of a route that nests other routes.

## How I Tested These Changes

Haven't been able to repro the issue myself but you can see this in the
go/ux dashboard. Seems to happen for a particularly large org with slow
and expensive renders