[ui] Add a coordinator logs tab to the backfill detail page (#22409)
## Summary & Motivation

The backfill logs are implemented as a cursor-paginated resolver on
partitionBackfill. To get the latest logs, you have to start at the
first log (no cursor) and paginate to get to the last log. When you
reach the last log, it returns `hasMore: false`, and we switch to
polling that final cursor for additional logs and advancing it as they
come in.

This does mean that showing the "tail" of the logs requires loading them
all, but we're not super worried about this -- because the backend
implementation stores the logs as files on S3, the underlying
implementation doesn't support a "last N" retrieval.

I created a new query hook `useCursorAccumulatedQuery` which is similar
to `useCursorPaginatedQuery` but effectively auto-pages and accumulates
the data rather than returning a single page at a time.

Misc:

- I also added the refresh countdown components to all the backfill tabs

Important note:

- This feature is shipping in OSS this week but has not yet shipped -
backfill logs are currently only available in cloud when an org setting
is enabled.

## How I Tested These Changes

- Because the OSS backend for this is still in-flight I wrote a test to
cover the front-end logic

---------

Co-authored-by: bengotow <bgotow@elementl.com>